transforming_waste_management/flask_project/requirement.txt
Flask
tensorflow
numpy
Pillow
werkzeug